requirements.txt
numpy>=1.16.2
pillow>=5.2.0
scipy>=1.4.1
opencv-python
azure-storage-blob>=12.0.0
azure-common
nbstripout
pytest-azurepipelines
pytest-cov
pytest
requests
twine
wheel
